fix: Publish Factory workspace changes through Eve
