+++ apps/factory/agent/lib/daily-example-maintenance.ts
@@ -10,5 +10,5 @@ export function fxExampleMaintenancePrompt(
 ): string {
   return `Maintain only the examples/${example} example; do not inspect or update any other example. Audit and update its stale dependencies, package-manager pin, Node engine, README instructions, versioned references, and turbo.json tasks. Use exact latest stable versions, apply required best-practice migrations, regenerate its lockfile with its declared package manager, and run every relevant non-persistent validation task. Fix validation failures rather than stopping at an audit report.
 
-When changes exist, create a branch named agents/examples-${example}-fx-${sessionId}, commit with the title "chore: Update ${example} example", run \`gh auth setup-git\`, push it, and open a draft pull request against vercel/turborepo. Reuse an existing open pull request for that branch rather than creating a duplicate. Summarize the changes and validation results in the body. GH_TOKEN contains a non-secret placeholder; the sandbox firewall supplies GitHub authorization. If there are no changes, finish without creating a pull request.`;
+When changes exist, leave them uncommitted and run \`factory-create-pr --branch agents/examples-${example}-fx-${sessionId} --title "chore: Update ${example} example" --body "<summary and validation results>"\`. Factory creates the verified commit, branch, and draft pull request through Eve's GitHub credentials. Never run git commit, git push, gh auth setup-git, or gh pr create. If there are no changes, finish without creating a pull request.`;
 }
